[AIRFLOW-5807] Move SFTP from contrib to providers. (#6464)
* [AIRFLOW-5807] Move SFTP from contrib to core
