+++ docs/autoapi_templates/index.rst
@@ -76,6 +76,8 @@ All operators are in the following packages:
 
   airflow/providers/amazon/aws/sensors/index
 
+  airflow/providers/apache/cassandra/sensors/index
+
   airflow/providers/google/cloud/operators/index
 
   airflow/providers/google/cloud/sensors/index
@@ -84,7 +86,9 @@ All operators are in the following packages:
 
   airflow/providers/google/marketing_platform/sensors/index
 
-  airflow/providers/apache/cassandra/sensors/index
+  airflow/providers/sftp/operators/index
+
+  airflow/providers/sftp/sensors/index
 
 Hooks
 -----
@@ -116,6 +120,8 @@ All hooks are in the following packages:
   airflow/providers/google/cloud/hooks/index
 
   airflow/providers/apache/cassandra/hooks/index
+
+  airflow/providers/sftp/hooks/index
 
 Executors
 ---------
