+++ docs/operators-and-hooks-ref.rst
@@ -1246,9 +1246,9 @@ communication protocols or interface.
 
    * - `SSH File Transfer Protocol (SFTP) <https://tools.ietf.org/wg/secsh/draft-ietf-secsh-filexfer/>`__
      -
-     - :mod:`airflow.contrib.hooks.sftp_hook`
-     - :mod:`airflow.contrib.operators.sftp_operator`
-     - :mod:`airflow.contrib.sensors.sftp_sensor`
+     - :mod:`airflow.providers.sftp.hooks.sftp_hook`
+     - :mod:`airflow.providers.sftp.operators.sftp_operator`
+     - :mod:`airflow.providers.sftp.sensors.sftp_sensor`
 
    * - `Secure Shell (SSH) <https://tools.ietf.org/html/rfc4251>`__
      -
